Additional functions 2 and 3
I left blur as it was from the last commmit, it works although I wanted
to change the code but didn't get the chance, so I'm leaving it as it
is. Originally Additional function 3 was resize but there was a problem
where the original pixels still displayed after the resize. So I changed
this function to be a desaturate image function instead.